# apiceclinic-module/clinic-api/generated-auth/src/main/kotlin/com/clinic/model/LoginResponse.kt
/**
 *
 * Please note:
 * This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit this file manually.
 *
 */

@file:Suppress(
    "ArrayInDataClass",
    "EnumEntryName",
    "RemoveRedundantQualifierName",
    "UnusedImport"
)

package com.clinic.model


import com.fasterxml.jackson.annotation.JsonProperty

/**
 * Risposta al login con JWT token
 *
 * @param accessToken JWT token da usare nelle richieste autenticate
 * @param tokenType Tipo di token (sempre \"Bearer\")
 * @param username Username dell'utente autenticato
 * @param role Ruolo dell'utente (ROLE_USER o ROLE_ADMIN)
 * @param userId ID univoco dell'utente autenticato
 * @param patientId ID del profilo paziente collegato (null per admin)
 */


data class LoginResponse (

    /* JWT token da usare nelle richieste autenticate */
    @field:JsonProperty("accessToken")
    val accessToken: kotlin.String,

    /* Tipo di token (sempre \"Bearer\") */
    @field:JsonProperty("tokenType")
    val tokenType: kotlin.String,

    /* Username dell'utente autenticato */
    @field:JsonProperty("username")
    val username: kotlin.String,

    /* Ruolo dell'utente (ROLE_USER o ROLE_ADMIN) */
    @field:JsonProperty("role")
    val role: kotlin.String,

    /* ID univoco dell'utente autenticato */
    @field:JsonProperty("userId")
    val userId: kotlin.Long,

    /* ID del profilo paziente collegato (null per admin) */
    @field:JsonProperty("patientId")
    val patientId: kotlin.Long? = null

)